仮の街を導入
仮の街をプレハブとして導入。
