--- Assets/City_Assets.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 88377c79a28b03d4aa1c4b2043be4409
-folderAsset: yes
-timeCreated: 1510040022
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
