+++ Assets/Map/City.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: ac09f147052b03841b9ebabcef3e293f
+timeCreated: 1510644892
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
