--- Assets/Resources/Prefabs/Game/CityAdventureMap_1.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: cd49faadf2c8d964f84946eb7c9d09ee
-timeCreated: 1509434283
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
