+++ Assets/Resources/Prefabs/Game/CitySpawnManager.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 05cf89f3cf6fc724896a1edc82eceed4
+timeCreated: 1508390893
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
