--- Assets/Resources/Prefabs/Game/SpawnManager.prefab
@@ -1,68 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1001 &100100000
-Prefab:
-  m_ObjectHideFlags: 1
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications: []
-    m_RemovedComponents: []
-  m_ParentPrefab: {fileID: 0}
-  m_RootGameObject: {fileID: 1648512796981284}
-  m_IsPrefabParent: 1
---- !u!1 &1648512796981284
-GameObject:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  serializedVersion: 5
-  m_Component:
-  - component: {fileID: 4919885139050692}
-  - component: {fileID: 114976934096940558}
-  m_Layer: 0
-  m_Name: SpawnManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &4919885139050692
-Transform:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1648512796981284}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &114976934096940558
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1648512796981284}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b3626971b0fcdc841bf1a8abad682f5e, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  spawnPointPrefab: {fileID: 1419229282505804, guid: 90696ef91b711c346b5ab80b425164b5,
-    type: 2}
-  spawnPoint:
-  - {x: -150, y: 1, z: 150}
-  - {x: 0, y: 1, z: 150}
-  - {x: 150, y: 1, z: 150}
-  - {x: -150, y: 1, z: 0}
-  - {x: 0, y: 1, z: 0}
-  - {x: 150, y: 1, z: 0}
-  - {x: -150, y: 1, z: -150}
-  - {x: 0, y: 1, z: -150}
-  - {x: 150, y: 1, z: -150}
-  pearPassengerNum: 2
-  smallPassengerNum: 3
-  bigPassengerNum: 5
