--- Assets/Resources/Prefabs/Game/SpawnManager.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 05cf89f3cf6fc724896a1edc82eceed4
-timeCreated: 1508390893
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
