+++ Assets/Resources/Prefabs/Result.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 73d94452326fcab4191f7c644b41db5b
+folderAsset: yes
+timeCreated: 1510476607
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
