+++ Assets/Resources/Prefabs/Title.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 146ad6d7e0a23704ca534e78e247da17
+folderAsset: yes
+timeCreated: 1510476599
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
