+++ Assets/Resources/Prefabs/Tutorial.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 25dacc2a45c9c6a4fbce35c887aa1c35
+folderAsset: yes
+timeCreated: 1510476614
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
