--- Assets/Scripts/SpawnManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b3626971b0fcdc841bf1a8abad682f5e
-timeCreated: 1507624993
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
